# ApnaStayBackEnd/docs/swagger.json
  {
    "openapi": "3.0.3",
    "info": {
      "title": "HomeFinity Backend API",
      "description": "API for HomeFinity frontend.\n\n**Required headers for protected endpoints:**\n- **Authorization** (mandatory): `Authorization: Bearer <JWT>` — JWT from POST /api/auth/public/signin or POST /api/auth/public/phone/verify-and-login.\n- **X-XSRF-TOKEN** (mandatory): CSRF token from GET /api/csrf-token.\n\n**Required header for requests with body (POST/PUT):**\n- **Content-Type** (mandatory): `Content-Type: application/json`.\n\nPublic endpoints (signin, signup, phone send-code, phone verify-and-login, forgot-password, reset-password, verify-2fa-login, csrf-token, health-check) do not require Authorization or X-XSRF-TOKEN.",
      "version": "0.2.0"
    },
    "servers": [
      {
        "url": "/",
        "description": "Relative to deployed base URL"
      }
    ],
...
    "security": [
      {
        "bearerAuth": [],
        "csrfHeader": []
      }
    ],
    "paths": {
...
      "/api/auth/public/signin": {
        "post": {
          "tags": ["Auth"],
          "summary": "Sign in",
          "description": "Authenticate and receive JWT. No Authorization or X-XSRF-TOKEN required. Content-Type: application/json required (request body).",
          "operationId": "signin",
          "security": [],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/LoginRequest" }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Login successful",
              "content": {
                "application/json": {
                  "schema": { "$ref": "#/components/schemas/LoginResponse" }
                }
              }
            },
            "404": {
              "description": "Bad credentials",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "message": { "type": "string", "example": "Bad credentials" },
                      "status": { "type": "boolean", "example": false }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/api/auth/public/signup": {
        "post": {
          "tags": ["Auth"],
          "summary": "Sign up",
          "description": "Register a new user. No Authorization or X-XSRF-TOKEN required. Content-Type: application/json required (request body).",
          "operationId": "signup",
          "security": [],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SignupRequest" }
              }
            }
          },
          "responses": {
            "200": {
              "description": "User registered successfully",
              "content": {
                "application/json": {
                  "schema": { "$ref": "#/components/schemas/ApiResponseMessage" }
                }
              }
            },
            "400": {
              "description": "Validation error or username/email/phone already taken",
              "content": {
                "application/json": {
                  "schema": { "$ref": "#/components/schemas/MessageResponse" }
                }
              }
            }
          }
        }
      },
      "/api/auth/public/phone/send-code": {
        "post": {
          "tags": ["Auth"],
          "summary": "Send phone verification code (OTP)",
          "description": "Send a verification code via SMS to the given phone number. Use this as step 1 before phone login or phone signup. No Authorization or X-XSRF-TOKEN required.",
          "operationId": "sendPhoneVerificationCode",
          "security": [],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PhoneVerificationRequest" }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Verification code sent successfully",
              "content": {
                "application/json": {
                  "schema": { "$ref": "#/components/schemas/ApiResponseMessage" }
                }
              }
            },
            "400": {
              "description": "Phone number is required or invalid",
              "content": {
                "application/json": {
                  "schema": { "$ref": "#/components/schemas/ApiResponseError" }
                }
              }
            },
            "500": {
              "description": "Failed to send verification code",
              "content": {
                "application/json": {
                  "schema": { "$ref": "#/components/schemas/ApiResponseError" }
                }
              }
            }
          }
        }
      },
      "/api/auth/public/phone/verify-and-login": {
        "post": {
          "tags": ["Auth"],
          "summary": "Verify OTP and login or sign up (phone)",
          "description": "Verify the OTP sent to the phone number. If user exists, logs in; otherwise creates a new user (phone signup) and returns JWT. No Authorization or X-XSRF-TOKEN required.",
          "operationId": "phoneVerifyAndLogin",
          "security": [],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PhoneLoginRequest" }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Login/signup successful; returns JWT and user info",
              "content": {
                "application/json": {
                  "schema": { "$ref": "#/components/schemas/LoginResponse" }
                }
              }
            },
            "400": {
              "description": "Phone number is required or invalid",
              "content": {
                "application/json": {
                  "schema": { "$ref": "#/components/schemas/ApiResponseError" }
                }
              }
            },
            "401": {
              "description": "Invalid or expired verification code",
              "content": {
                "application/json": {
                  "schema": { "$ref": "#/components/schemas/ApiResponseError" }
                }
              }
            },
            "500": {
              "description": "Error during phone authentication",
              "content": {
                "application/json": {
                  "schema": { "$ref": "#/components/schemas/ApiResponseError" }
                }
              }
            }
          }
        }
      },
      "/api/auth/user": {
        "get": {
          "tags": ["Auth"],
          "summary": "Get current user details",
          "description": "Returns the authenticated user's profile. Requires Authorization and X-XSRF-TOKEN (mandatory).",
          "operationId": "getUserDetails",
"parameters": [
          { "$ref": "#/components/parameters/XsrfTokenHeader" }
        ],
        "responses": {
          "200": {
            "description": "User details",
              "content": {
                "application/json": {
                  "schema": { "$ref": "#/components/schemas/ApiResponseUserInfo" }
                }
              }
            },
            "401": { "description": "Unauthorized" }
          }
        }
      },
      "/api/auth/username": {
        "get": {
          "tags": ["Auth"],
          "summary": "Get current username",
          "description": "Returns the authenticated user's username. Requires Authorization and X-XSRF-TOKEN (mandatory).",
          "operationId": "getUsername",
"parameters": [
          { "$ref": "#/components/parameters/XsrfTokenHeader" }
        ],
        "responses": {
          "200": {
            "description": "Username or error",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "success": { "type": "boolean" },
                      "message": { "type": "string" },
                      "timestamp": { "type": "string", "format": "date-time" },
                      "data": { "type": "string", "description": "Username when success is true" }
                    }
                  }
                }
              }
            },
            "401": { "description": "Unauthorized" }
          }
        }
      },
...
      "/api/admin/user/{id}": {
        "get": {
          "tags": ["Admin"],
          "summary": "Get user by ID",
          "description": "Returns a user by ID (admin only). Requires Authorization and X-XSRF-TOKEN (mandatory).",
          "operationId": "getUserById",
          "parameters": [
            { "$ref": "#/components/parameters/XsrfTokenHeader" },
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": { "type": "integer", "format": "int64" }
            }
          ],
          "responses": {
            "200": {
              "description": "User details",
              "content": {
                "application/json": {
                  "schema": { "$ref": "#/components/schemas/ApiResponseUserDTO" }
                }
              }
            },
            "401": { "description": "Unauthorized" },
            "403": { "description": "Forbidden" },
            "404": { "description": "User not found" }
          }
        }
      },
...
      "/api/auth/public/forgot-password": {
        "post": {
          "tags": ["Auth"],
          "summary": "Forgot password",
          "description": "Send password reset email. No Authorization or X-XSRF-TOKEN required.",
          "operationId": "forgotPassword",
          "security": [],
          "parameters": [
            { "name": "email", "in": "query", "required": true, "schema": { "type": "string", "format": "email" } }
          ],
          "responses": {
            "200": {
              "description": "Password reset email sent",
              "content": {
                "application/json": {
                  "schema": { "$ref": "#/components/schemas/ApiResponseMessage" }
                }
              }
            },
            "500": {
              "description": "Failed to send email",
              "content": {
                "application/json": {
                  "schema": { "$ref": "#/components/schemas/ApiResponseError" }
                }
              }
            }
          }
        }
      },
      "/api/auth/public/reset-password": {
        "post": {
          "tags": ["Auth"],
          "summary": "Reset password",
          "description": "Reset password using token from forgot-password email. No Authorization or X-XSRF-TOKEN required.",
          "operationId": "resetPassword",
          "security": [],
          "parameters": [
            { "name": "token", "in": "query", "required": true, "schema": { "type": "string" } },
            { "name": "newPassword", "in": "query", "required": true, "schema": { "type": "string" } }
          ],
          "responses": {
            "200": {
              "description": "Password reset successful",
              "content": {
                "application/json": {
                  "schema": { "$ref": "#/components/schemas/ApiResponseMessage" }
                }
              }
            },
            "400": {
              "description": "Invalid or expired token",
              "content": {
                "application/json": {
                  "schema": { "$ref": "#/components/schemas/ApiResponseError" }
                }
              }
            }
          }
        }
      },
      "/api/auth/enable-2fa": {
        "post": {
          "tags": ["Auth"],
          "summary": "Enable 2FA",
          "description": "Generate 2FA secret and return QR code URL. Requires Authorization and X-XSRF-TOKEN.",
          "operationId": "enable2FA",
          "parameters": [
            { "$ref": "#/components/parameters/XsrfTokenHeader" }
          ],
          "responses": {
            "200": {
              "description": "QR code URL for 2FA setup",
              "content": {
                "text/plain": {
                  "schema": { "type": "string", "description": "QR code URL" }
                }
              }
            },
            "401": { "description": "Unauthorized" }
          }
        }
      },
      "/api/auth/disable-2fa": {
        "post": {
          "tags": ["Auth"],
          "summary": "Disable 2FA",
          "description": "Disable two-factor authentication. Requires Authorization and X-XSRF-TOKEN.",
          "operationId": "disable2FA",
          "parameters": [
            { "$ref": "#/components/parameters/XsrfTokenHeader" }
          ],
          "responses": {
            "200": {
              "description": "2FA disabled",
              "content": {
                "text/plain": {
                  "schema": { "type": "string", "example": "2FA disabled" }
                }
              }
            },
            "401": { "description": "Unauthorized" }
          }
        }
      },
      "/api/auth/verify-2fa": {
        "post": {
          "tags": ["Auth"],
          "summary": "Verify 2FA code",
          "description": "Verify 2FA code to complete 2FA setup. Requires Authorization and X-XSRF-TOKEN.",
          "operationId": "verify2FA",
          "parameters": [
            { "$ref": "#/components/parameters/XsrfTokenHeader" },
            { "name": "code", "in": "query", "required": true, "schema": { "type": "integer", "format": "int32" } }
          ],
          "responses": {
            "200": {
              "description": "2FA verified and enabled",
              "content": {
                "text/plain": {
                  "schema": { "type": "string", "example": "2FA verified" }
                }
              }
            },
            "401": {
              "description": "Invalid 2FA code",
              "content": {
                "text/plain": {
                  "schema": { "type": "string", "example": "Invalid 2FA code" }
                }
              }
            }
          }
        }
      },
      "/api/auth/user/2fa-status": {
        "get": {
          "tags": ["Auth"],
          "summary": "Get 2FA status",
          "description": "Returns whether 2FA is enabled for the current user. Requires Authorization and X-XSRF-TOKEN.",
          "operationId": "get2FAStatus",
          "parameters": [
            { "$ref": "#/components/parameters/XsrfTokenHeader" }
          ],
          "responses": {
            "200": {
              "description": "2FA status",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "is2faEnabled": { "type": "boolean" }
                    }
                  }
                }
              }
            },
            "401": { "description": "Unauthorized" },
            "404": { "description": "User not found" }
          }
        }
      },
      "/api/auth/public/verify-2fa-login": {
        "post": {
          "tags": ["Auth"],
          "summary": "Verify 2FA during login",
          "description": "Verify 2FA code during login (after signin returns JWT). No Authorization required.",
          "operationId": "verify2FALogin",
          "security": [],
          "parameters": [
            { "name": "code", "in": "query", "required": true, "schema": { "type": "integer", "format": "int32" } },
            { "name": "jwtToken", "in": "query", "required": true, "schema": { "type": "string" } }
          ],
          "responses": {
            "200": {
              "description": "2FA verified",
              "content": {
                "text/plain": {
                  "schema": { "type": "string", "example": "2FA Verified" }
                }
              }
            },
            "401": {
              "description": "Invalid 2FA code",
              "content": {
                "text/plain": {
                  "schema": { "type": "string", "example": "Invalid 2FA code" }
                }
              }
            }
          }
        }
      },
...
    },
    "components": {
...
      "securitySchemes": {
        "bearerAuth": {
          "type": "http",
          "scheme": "bearer",
          "bearerFormat": "JWT",
          "description": "JWT token from /api/auth/public/signin. Send as: Authorization: Bearer <token>"
        },
        "csrfHeader": {
          "type": "apiKey",
          "in": "header",
          "name": "X-XSRF-TOKEN",
          "description": "CSRF token from GET /api/csrf-token. Use the header name returned by that endpoint (e.g. X-XSRF-TOKEN) and the token value. Required for state-changing requests and optionally for GETs when cookies are used."
        }
      },
      "schemas": {
...
        "LoginRequest": {
          "type": "object",
          "required": ["username", "password"],
          "properties": {
            "username": { "type": "string" },
            "password": { "type": "string" }
          }
        },
...
      }
    }
  }